+++ actors/user/src/main/java/org/sunbird/user/actors/UserLoginActor.java
@@ -1,0 +1,54 @@
+package org.sunbird.user.actors;
+
+import org.sunbird.actor.router.ActorConfig;
+import org.sunbird.common.exception.ProjectCommonException;
+import org.sunbird.common.models.response.Response;
+import org.sunbird.common.models.util.ActorOperations;
+import org.sunbird.common.models.util.JsonKey;
+import org.sunbird.common.models.util.ProjectLogger;
+import org.sunbird.common.models.util.PropertiesCache;
+import org.sunbird.common.request.ExecutionContext;
+import org.sunbird.common.request.Request;
+import org.sunbird.common.responsecode.ResponseCode;
+import org.sunbird.learner.util.Util;
+
+@ActorConfig(
+	    tasks = {"updateUserLoginTime"},
+	    asyncTasks = {})
+public class UserLoginActor extends UserBaseActor {
+
+  @Override
+  public void onReceive(Request request) throws Throwable {
+    Util.initializeContext(request, JsonKey.USER);
+    // set request id fto thread loacl...
+    ExecutionContext.setRequestId(request.getRequestId());
+    String operation = request.getOperation();
+    if (operation.equalsIgnoreCase(ActorOperations.USER_CURRENT_LOGIN.getValue())) {
+      updateUserLoginTime(request);
+    } else {
+      ProjectLogger.log("UNSUPPORTED OPERATION");
+      ProjectCommonException exception =
+          new ProjectCommonException(
+              ResponseCode.invalidOperationName.getErrorCode(),
+              ResponseCode.invalidOperationName.getErrorMessage(),
+              ResponseCode.CLIENT_ERROR.getResponseCode());
+      sender().tell(exception, self());
+    }
+  }
+
+  /**
+   * This method will update user current login time in keycloak
+   *
+   * @param actorMessage Request
+   */
+  private void updateUserLoginTime(Request actorMessage) {
+    String userId = (String) actorMessage.getRequest().get(JsonKey.USER_ID);
+    Response response = new Response();
+    response.put(JsonKey.RESPONSE, JsonKey.SUCCESS);
+    sender().tell(response, self());
+    if (Boolean.parseBoolean(PropertiesCache.getInstance().getProperty(JsonKey.IS_SSO_ENABLED))) {
+      boolean addedResponse = getSsoManager().addUserLoginTime(userId);
+      ProjectLogger.log("user login time added response is ==" + addedResponse);
+    }
+  }
+}
